requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
tavily-python>=0.5.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
linkedin-api>=2.3.0
promptquality>=2.0.0
streamlit>=1.39.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.39.0
